feat: wire down command to use removeEnvFiles with external path safety
